+++ common/src/main/kotlin/net/stellarica/common/coordinate/RelativeBlockPosition.kt
@@ -3,7 +3,6 @@ package net.stellarica.common.coordinate
 import kotlinx.serialization.Serializable
 import net.minecraft.core.BlockPos
 import net.minecraft.core.Direction
-import net.stellarica.common.util.*
 
 
 /**
@@ -15,14 +14,14 @@ data class RelativeBlockPosition(
     val y: Int,
     val z: Int
 ) {
-    fun getBlockPosition(origin: BlockPos, direction: Direction): BlockPos {
+    fun getGlobalPosition(origin: BlockPosition, direction: Direction): BlockPosition {
         return when (direction) {
-            Direction.NORTH -> BlockPos(this.x, this.y, this.z)
-            Direction.EAST -> BlockPos(-this.z, this.y, this.x)
-            Direction.SOUTH -> BlockPos(-this.x, this.y, -this.z)
-            Direction.WEST -> BlockPos(this.z, this.y, -this.x)
+            Direction.NORTH -> BlockPosition(this.x, this.y, this.z)
+            Direction.EAST -> BlockPosition(-this.z, this.y, this.x)
+            Direction.SOUTH -> BlockPosition(-this.x, this.y, -this.z)
+            Direction.WEST -> BlockPosition(this.z, this.y, -this.x)
             else -> throw IllegalArgumentException()
-        }.offset(origin)
+        } + origin
     }
 
     operator fun plus(other: RelativeBlockPosition) = RelativeBlockPosition(x + other.x, y + other.y, z + other.z)
