+++ server/src/main/kotlin/net/stellarica/server/craft/BlockContainer.kt
@@ -1,0 +1,18 @@
+package net.stellarica.server.craft
+
+import net.minecraft.core.Direction
+import net.stellarica.common.coordinate.BlockPosition
+import net.stellarica.common.coordinate.RelativeBlockPosition
+
+interface BlockContainer {
+    val origin: BlockPosition
+    val orientation: Direction
+
+    fun getGlobalPos(pos: RelativeBlockPosition): BlockPosition {
+        return pos.getGlobalPosition(origin, orientation)
+    }
+
+    fun getRelativePos(pos: BlockPosition): RelativeBlockPosition {
+        return pos.getAsRelative(origin, orientation)
+    }
+}
