+++ server/src/main/kotlin/net/stellarica/server/multiblock/MultiblockInstance.kt
@@ -3,15 +3,13 @@ package net.stellarica.server.multiblock
 import kotlinx.serialization.Serializable
 import net.minecraft.core.BlockPos
 import net.minecraft.core.Direction
-import net.stellarica.common.util.OriginRelative
-import net.stellarica.common.serializer.BlockPosSerializer
+import net.stellarica.common.coordinate.BlockPosition
 import net.stellarica.common.serializer.DirectionSerializer
 import net.stellarica.common.serializer.UUIDSerializer
+import net.stellarica.server.craft.BlockContainer
 import net.stellarica.server.multiblock.data.MultiblockData
 import net.stellarica.server.multiblock.type.MultiblockTypeSerializer
-import net.stellarica.server.util.extension.toLocation
 import net.stellarica.server.serializer.BukkitWorldSerializer
-import org.bukkit.Chunk
 import org.bukkit.World
 import java.util.UUID
 
@@ -19,8 +17,7 @@ import java.util.UUID
 class MultiblockInstance(
 		@Serializable(with = UUIDSerializer::class)
 		val id: UUID,
-		@Serializable(with = BlockPosSerializer::class)
-		val origin: BlockPos,
+		val origin: BlockPosition,
 		@Serializable(with = BukkitWorldSerializer::class)
 		val world: World,
 		@Serializable(with = DirectionSerializer::class)
@@ -28,19 +25,9 @@ class MultiblockInstance(
 		@Serializable(with = MultiblockTypeSerializer::class)
 		val type: MultiblockType,
 		val data: MultiblockData
-) {
-	val chunk: Chunk
-		get() = world.getChunkAt(origin.toLocation(world))
-
+): BlockContainer {
 
 	fun validate() = type.validate(direction, origin, world)
-
-	/** @return the world location of [position] */
-	fun getLocation(position: OriginRelative) = position.getBlockPos(origin, direction)
-
-	/** @return the origin relative position of [loc] */
-	fun getOriginRelative(loc: BlockPos) = OriginRelative.getOriginRelative(loc, this.origin, this.direction)
-
 
 	/**
 	 * @return whether this contains a blocks at [loc].
